add support for differently sized slides
The newly added 'autoSlideSize' option will allow the slides to have the same height and variable width if they are horizontally oriented, or same width and variable height if they are vertically oriented. Some code refactoring was done in the core code and a few modules in order to accommodate this possibility.